tests/wpt/tests/html/webappapis/dynamic-markup-insertion/document-write/035.html

<!doctype html>
<title>document.write</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
t.step(function() {
  var s = "<svg><!";
  for (var i=0; i<s.length; i++) {
    document.write(s[i]);
  }
});
</script>[CDATA[Filler Text]]></svg>
<script>
t.step(function() {
  assert_equals(document.body.childNodes[0].localName, "svg");
  assert_equals(document.body.childNodes[0].textContent, "Filler Text");
}
);
t.done();
</script>
<div id="log"></div>
